Code Move, Text Wrapping, Control Searching
Move the drawing code to its own file to make things more tidy.

Control searching is part of text wrapping.  Controls now have a Name member that is a required unique identfier used for searching the window arrays  for a particular control using corresponding Find methods.

This is needed so that controls can have references to each other, especially ZShapes.