Borders for Standard Buttons
Discovered what is either a bug or just, hopefully, an oversight in ZScript: lines are not clipped off by SetClipRect.  I've made a post about this, but have yet to receive an answer one way or another, so I'm going to code around it for the meantime.

I also forgot to take the right and bottom edges of the window into account, which is fixed now.