Passive Button Interactions
Passing the cursor information to the windows is done.

Passive Button Interactions - called GibZoning - is done.  Buttons now highlight on mouse-over, if they have supporting graphics, and if they are not covered by another window.

I haven't decided if there should be any kind of direct access to the the handler's window stack, so currently the only way to get any information on a window is through a WindowStats class, which supplies some basic information right now; the full window could be supported making this a read-only access system.  Users can get the WindowStats of a window by either it's index in the stack or directly by name.

Also added a GetStackIndex method, which will call the array's Find method.  This does require having the window in the first place.