Tweak to button GibZoning & Priority Switching
Needed to check if the button is entirely covered or not before looking for partial coverage.

Windows that are invalid for the given consoleplayer are ignored in GibZoning checks - the window might become valid so it is allowed to persist in the window stack.

Priority switching works - the check is pretty stupid simple, run the stack in reverse, whoever passes the click check is the winner and gets highest priority on the next draw call.