UIToggle, Controls moved to Play Scope, Handler Fixes
UIToggle allows a window that is being created for the first time to immediately toggle the mouse.

ZSWin_ControlBase and descendents (Buttons, Text, Shapes) have been moved the the Play scope.  This allows ZSWindow-level acces, since the events send the window, so controls can do stuff to the windows.  The ZText method, GetAlignment was moved to UI as a result, it's a UI calculation method anyway.

Closing a window didn't actually work, priority wasn't taken into account, nor was null indices in the window stack.  This is fixed, windows can come and go as needed.

Console window is just created by the handler now, thought I needed an actor so I was hijacking the LostSoul - too bad I was hoping to create an exploding LostSoul appear in front of the player that barfed the console window onto the screen.