Event Extensions, QuikClose Update
UiTick and NetworkProcess have been extended to ZObjects to allow objects to send and receive their own net commands.  This is in preparation for a TextBox control which will accept keyboard input.

QuikClose also requires modification in order to play nice with keyboard input.  The Esc key and Tilde key are also supported.