Font Serialization Fix for ZText
The ZText class no longer contains an actual Font reference and instead just stores the name of the font; this member, TextFont, is a name, not a string.  All usage of fonts needs to use the global Font class methods.  ZSWin objects cannot contain Font class members - they are not serializable.