Command System Maintenance & Focusing
Patched, it's not the best solution, the command system so that it catches empty, invalid, or just broken commands.  This is done for the event system, not objects - still need to see if it's necessary to do the same for them, most commands are internal.

Separated focusing controls from control priority.  Priority works just like window priority, allowing controls to switch draw order based on interaction.  Focus is more selective, only one control may have focus, for what intent is controlled by the control.  Focusing is done the same as priority switching, cursor location validation and response in a mouse event.