Synchronization Fix - Still not fixed - Documentation Update - Dev Build Revision
Partially fixed the synchronization problem.  Adding a window to the stack is simply a job for the window itself, so users don't interact with that anymore.  This is causing multiplayer crashes though, so still work to be done - not 100% sure what right now.

Fixed a bunch of the documentation in demo files as a result, a lot of that was garbage from old builds.

Reverted the changes that forced users to use a recent dev build - it was literally 1 line of code - this can be held off until an official release supports the "abstract function".