v0.4.1 Update - Bug Fixes & USDF Dialog Control
Event System has been returned to using exclusively ZScript types for objects.  The BST for the "allZObjects" list has been removed and it is now a dynamic array.  Further supporting changes made to the Event System.

Noticed the "allZObjects" array was getting duplicates during object creation - this had to do with network events and how initialization is timed, which breaks the system's enforcement of unique names for objects in multiplayer.  This is fixed through a lockout mechanism for that particular global event.

(Incomplete) Exported USDF - or an equivalent of - to ZScript.  This control take advantage of the majority of existing ZScript Windows features to create the visual objects, buttons, text, etc., which allows the control to be focused on the internal operation of USDF.  Full documentation to come; most USDF features are supported, largely unimplemented due to bug fixing: Pages, IfItems, Choices, and Costs have a representation. Features can be added as well.

Currently chasing a multiplayer bug with the mouse cursor that is breaking interactivity for anyone but the first player.