v0.4.1.1 Multiplayer Bugfix and ZConversation Update
Implemented third, and correct, solution to a bug that jumped between single and multiplayer depending on my chosen solution - basically the global object array was getting duplicates when it should never, ever have duplicates, there's code to prevent that but it was failing due to access racing.  So the Event System needed to handle the "add the object to the global list" command more precisely, and how that command is sent needed changed.

The ZConversation can now wait a defined number to ticks before changing pages to allow responses to be displayed.  Choice buttons are fixed to get the right choice index they correspond to; I was being lazy about that originally.