v0.4.1.3 Update Partial USDF DropItem Support
It's almost there - I'm out of time for the moment so just posting the existing code.  Right now this does not work, the order of events is wrong.  The ZConversation needs to access the event handler directly and skip net events to create the EventDataPacket, however the net command method can remain.

What is an EventDataPacket?  A generic data packet that can be processed by any of the handler events.  Right now the PoC works with WorldThingDied.