v0.4.1.3 Update - Event System, EventDataPacket
This updates the event system so that EventDataPackets can be created directly; this is dangerous but necessary due to the order of operations in some circumstances.

Testing of item dropping using WorldThingDied has demonstrated this works in multiplayer as well, and functions as the game intends.

Other noted bugs: the cursor system is not getting the correct player information in multiplayer.  Solution currently unknown. NPCs killed by regular enemies (not players) will crash the game.  Again solution currently unknown.